Changed logic
Program now checks whether the URL contains the word crai. If so we cut the
substring differently. I also changed the naming from FILENAME to OBJECTNAME,
which is the name the file will show in the Google bucket UI.
Further, I removed the curly braces in the curl and gsutil command as that
took it as part of the filename.