Only process one URL per core
The previous problem using multithreading was that the machine
(EC2 t2.2xlarge) started choking as it attempted to work on 432
URLs at the same time.
This change now works on blocks of the size of the number of
CPUs of the system, and waits until each process has finished before
starting a new iteration.