Using split command
In the effort to break the number of simultaneous processs down
I used the split command in conjunction with a line number that's
dynamically computed. In a subdirectory, the original file is broken
down into many files. The main loop then iterates over each of these
files. Each file contains the same number of URLs, and all URLs in
one file are processed simultaneously.